+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 ### __Breaking Changes__
 
 ### Added
+- Draw: Adds a new attribute semicolonCSVDelimiter to decide the semicolon is used as the delimiter for exported CSV file. The default value is true.
 
 ### Changed